requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python-headless>=4.5.0
numpy>=1.21.0
pillow>=8.0.0
gradio>=3.0.0
matplotlib>=3.3.0
scipy>=1.7.0
requests>=2.25.0
scikit-image>=0.18.0
torchmetrics>=0.7.0
